extract recommended products section
